src/db/main.py
# We will create the async engine as we are using the async db api

from sqlmodel import create_engine, SQLModel
from sqlmodel.ext.asyncio.session import AsyncSession

from sqlalchemy.ext.asyncio import AsyncEngine
from sqlalchemy.orm import sessionmaker
from src.config import Config

async_engine = AsyncEngine(
    create_engine(
    echo=True,
    url=Config.DATABASE_URL,
))


async def init_db(): 
    async with async_engine.begin() as connection: 
        from src.books.models import Book 
        
        await connection.run_sync(SQLModel.metadata.create_all)


async def get_session():
    Session = sessionmaker(
        bind=async_engine,
        class_=AsyncSession,
        expire_on_commit=False,
    )
    async with Session() as session:
        yield session
