requirements.txt
fastapi
fastapi[standard]
pydantic
pydantic-settings
asyncpg
sqlmodel
alembic
passlib
bcrypt
pyjwt
redis
fastapi-mail